ReviewCount and AvgReviewValue are now calculated during getOffer

+++ offer/src/main/java/org/shareio/backend/core/usecases/port/in/GetOwnerReviewCountUseCaseInterface.java
@@ -3,6 +3,6 @@ package org.shareio.backend.core.usecases.port.in;
 import java.util.UUID;
 
 public interface GetOwnerReviewCountUseCaseInterface {
-    Long getUserReviewCount(UUID ownerId);
+    Integer getUserReviewCount(UUID ownerId);
 
 }
